functions/requirements.txt
firebase-functions==0.2.0
firebase-admin==6.5.0
genkit
google-generativeai==0.7.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
PyMuPDF==1.24.5